chore(people): renumber past main's 0223, and split the admission rules out
Main took 0223 for retention authoring while this branch was in review, so the
migration moves to 0224 — two different migrations at one version applies one
and silently skips the other for ever.

The admission code also pushed domaintriage.go past the file-length cap, and it
is a second subject: the triage asks what a domain's site says, while admission
answers whether the domain may have a company at all — a question no amount of
crawling settles, which is why it outlives any single read.
